+++ traffic_ops/client/api_capability.go
@@ -23,10 +23,13 @@ import (
    limitations under the License.
 */
 
-// GetAPICapabilities retrieves all (or filtered) api_capability from Traffic Ops.
+// GetAPICapabilities will retrieve API Capabilities. In the event that no capability parameter
+// is supplied, it will return all existing. If a capability is supplied, it will return only
+// those with an exact match. Order may be specified to change the default sort order.
 func (to *Session) GetAPICapabilities(capability string, order string) (tc.APICapabilityResponse, ReqInf, error) {
 	var (
 		vals   = url.Values{}
+		path   = fmt.Sprintf("%s/api_capabilities", apiBase)
 		reqInf = ReqInf{CacheHitStatus: CacheHitStatusMiss}
 		resp   tc.APICapabilityResponse
 	)
@@ -39,7 +42,10 @@ func (to *Session) GetAPICapabilities(capability string, order string) (tc.APICa
 		vals.Set("orderby", order)
 	}
 
-	path := fmt.Sprintf("%s/api_capabilities?%s", apiBase, vals.Encode())
+	if len(vals) > 0 {
+		path = fmt.Sprintf("%s?%s", path, vals.Encode())
+	}
+
 	httpResp, remoteAddr, err := to.request(http.MethodGet, path, nil)
 	reqInf.RemoteAddr = remoteAddr